WIP: fortify: swap to _chk-ish intrinsics
First pass at converting the FORTIFY magics in the kernel to use _chk
intrinsics, rather than having all size checks inline.

Notes are included for which `_chk` functions clang/gcc (since 5.1)
recognize. If a compiler recognizes `__foo_chk`, the compiler is capable
of optimizing it to `foo` if the `__foo_chk` variant would just call
`foo()` anyway.

For clang in particular, these transformations are a win. For GCC, it's
_potentially_ better, but less certain than for clang.